fix: Move test data to folder ignored by translation checker
